+++ init_web_v2.go
@@ -240,6 +240,7 @@ func initWebV2Actions(app *App) {
 	m.Get("/v3/accounts/{id}/sales", handlers.GetSaleListForAccount)
 	m.Get("/v3/accounts/{id}/sales/{sale_id}", handlers.GetSaleForAccount)
 	m.Get("/v3/accounts/{id}/converted_balances/{asset_code}", handlers.GetConvertedBalances)
+	m.Get("/v3/accounts/{id}/balances_statistic/{asset_code}", handlers.GetBalanceStatistic)
 	m.Get("/v3/assets/{code}", handlers.GetAsset)
 	m.Get("/v3/assets", handlers.GetAssetList)
 	m.Get("/v3/balances/{id}", handlers.GetBalance)
@@ -333,8 +334,8 @@ func initWebV2Actions(app *App) {
 
 	m.Get("/v3/operations", handlers.GetOperations)
 
-	janus := app.config.Janus()
-	if err := janus.RegisterChi(m); err != nil {
+	cop := app.config.Cop()
+	if err := cop.RegisterChi(m); err != nil {
 		panic(errors.Wrap(err, "failed to register service"))
 	}
 }